Pataisiau ant neegzistuojancio sectiono kurima. Dabar jeigu yra sectionai by default is karto rodo pirmo postus (reikes padaryti vizualu indikatoriu kad esi ant to sectiono, kaip nors patamsinti ji liste), o jei nera sectionu add post button patikrina ar section id nera 0 nes jis defaultinis uzstatytas. Perkeliau api i PostAPI faila.
